Add support for authentication and user creation using external HTTP headers
An external proxy (such as a Single Sign On service) can authenticate
the user and then let the user to access the app only if it has been
authenticated. In this case the user information is passed in special
user defined HTTP headers which can be trusted.

This patch adds the option to use these headers to authenticate the
user and if not present, also to create the user with appropriate
default settings.